# docs/api/openapi-spec.yml
...
openapi: 3.1.0
info:
  title: LogPilot Agent API
  description: 'Cloud-native, autonomous log-intelligence agent: ingestion, search, chat, RCA, reports, proactive
    failure forecasting and autonomy policy.


    Authenticate with POST /api/v1/auth/login, then send `Authorization: Bearer <token>`. Real-time events and streamed
    chat use the WebSocket at /ws?token=<jwt> (not described by OpenAPI).'
  version: 2.0.0
paths:
...
  /api/v1/policy/tools:
    get:
      tags:
      - policy
      summary: List Policy
      operationId: list_policy_api_v1_policy_tools_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/policy/tools/{tool_name}:
    put:
      tags:
      - policy
      summary: Set Policy
      operationId: set_policy_api_v1_policy_tools__tool_name__put
      security:
      - HTTPBearer: []
      parameters:
      - name: tool_name
        in: path
        required: true
        schema:
          type: string
          title: Tool Name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - policy
      summary: Delete Override
      description: Remove an environment-specific override (the '*' row is the base policy and stays).
      operationId: delete_override_api_v1_policy_tools__tool_name__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: tool_name
        in: path
        required: true
        schema:
          type: string
          title: Tool Name
      - name: scope
        in: query
        required: true
        schema:
          type: string
          title: Scope
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/settings/forecasting:
    get:
      tags:
      - policy
      summary: Get Forecast Settings
      operationId: get_forecast_settings_api_v1_settings_forecasting_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
    put:
      tags:
      - policy
      summary: Put Forecast Settings
      operationId: put_forecast_settings_api_v1_settings_forecasting_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForecastSettings'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/settings/provider:
    get:
      tags:
      - policy
      summary: Provider Settings
      description: 'Model/provider settings (read-only here: provider credentials live in the deployment''s secret
        store).'
      operationId: provider_settings_api_v1_settings_provider_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/webhooks:
    get:
      tags:
      - policy
      summary: List Webhooks
      operationId: list_webhooks_api_v1_webhooks_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
    post:
      tags:
      - policy
      summary: Create Webhook
      operationId: create_webhook_api_v1_webhooks_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookBody'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/webhooks/{webhook_id}:
    delete:
      tags:
      - policy
      summary: Delete Webhook
      operationId: delete_webhook_api_v1_webhooks__webhook_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Webhook Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/webhooks/{webhook_id}/test:
    post:
      tags:
      - policy
      summary: Test Webhook
      operationId: test_webhook_api_v1_webhooks__webhook_id__test_post
      security:
      - HTTPBearer: []
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Webhook Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ForecastSettings:
      properties:
        warning_threshold:
          type: integer
          maximum: 99.0
          minimum: 1.0
          title: Warning Threshold
        critical_threshold:
          type: integer
          maximum: 100.0
          minimum: 2.0
          title: Critical Threshold
        interval_seconds:
          type: integer
          maximum: 3600.0
          minimum: 10.0
          title: Interval Seconds
        window_seconds:
          type: integer
          maximum: 600.0
          minimum: 10.0
          title: Window Seconds
          default: 60
      type: object
      required:
      - warning_threshold
      - critical_threshold
      - interval_seconds
      title: ForecastSettings
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PolicyBody:
      properties:
        scope:
          type: string
          title: Scope
          default: '*'
        tier:
          type: string
          enum:
          - read_only
          - autonomous_background
          - autonomous_policy_bounded
          - propose_only
          - autonomous_execution
          title: Tier
        min_confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Min Confidence
          default: 0.0
        requires_approval:
          type: boolean
          title: Requires Approval
          default: false
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - tier
      title: PolicyBody
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WebhookBody:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        url:
          type: string
          title: Url
        secret:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Secret
        events:
          items:
            type: string
          type: array
          title: Events
        min_level:
          type: string
          enum:
          - warning
          - critical
          title: Min Level
          default: warning
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - name
      - url
      title: WebhookBody
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://localhost:8000
  description: Local (docker compose)